# server/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "FastCampus API Docs",
    "description": "FastCampus API 문서입니다.",
    "version": "1.0.0"
  },
  "host": "localhost:8000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
    "/api/v1/user/": {
      "post": {
        "tags": [
          "User"
        ],
        "description": "유저를 생성할 때 사용할 API Endpoint입니다.",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "description": "Patch User Data",
            "schema": {
              "$ref": "#/definitions/patchuser"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User successfully created.",
            "schema": {
              "$ref": "#/definitions/user"
            }
          }
        }
      }
    },
    "/api/v1/user/my": {
      "post": {
        "tags": [
          "User"
        ],
        "description": "내 정보를 조회합니다(로그인 상태여야 함).",
        "parameters": [],
        "responses": {
          "200": {
            "description": "User successfully created.",
            "schema": {
              "$ref": "#/definitions/user"
            }
          }
        }
      },
      "patch": {
        "tags": [
          "User"
        ],
        "description": "유저를 수정할 때 사용할 API Endpoint입니다.",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "description": "Patch User Data",
            "schema": {
              "$ref": "#/definitions/patchuser"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User successfully edited.",
            "schema": {
              "$ref": "#/definitions/user"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
    "user": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "example": "e879e3721010fa0c365f59151c6c3c878461cdc0"
        },
        "email": {
          "type": "string",
          "example": "remigailard80@korea.ac.kr"
        },
        "name": {
          "type": "string",
          "example": "김민석[ 학부재학 / 보건환경융합과학부 ]"
        },
        "phone": {},
        "email_certified": {
          "type": "number",
          "example": 0
        },
        "address": {},
        "organization": {},
        "company": {},
        "provider": {
          "type": "string",
          "example": "email"
        },
        "created_at": {
          "type": "string",
          "example": "2022-07-31 02:56:29"
        },
        "updated_at": {
          "type": "string",
          "example": "2022-07-31 02:56:29"
        }
      }
    },
...
    "patchuser": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "example": "e879e3721010fa0c365f59151c6c3c878461cdc0"
        },
        "phone": {
          "type": "string",
          "example": "01074616005"
        },
        "provider": {
          "type": "string",
          "example": "email"
        },
        "name": {
          "type": "string",
          "example": "안녕아녕"
        },
        "email": {
          "type": "string",
          "example": "remigailard801@naver.com"
        },
        "address": {
          "type": "string",
          "example": "경기도 성남시 수정구 ㅇㅇㅇ"
        },
        "organization": {
          "type": "string",
          "example": "Organization"
        },
        "company": {
          "type": "string",
          "example": "토양어선"
        }
      }
    }
  }
}